+++ Runtime/Scripts/RGOverlayMenu.cs
@@ -186,7 +186,7 @@ namespace RegressionGames
 
         public void StopBotInstance(long id)
         {
-            RGBotServerListener.GetInstance()?.HandleClientTeardown((uint) id);
+            RGBotServerListener.GetInstance()?.HandleClientTeardown(id);
         }
 
         /// <summary>
